add Get Report endpoint

+++ Backend-Test/src/BackendTest.Application/Models/ProductData.cs
@@ -1,4 +1,4 @@
 namespace BackendTest.Application.Models
 {
-    public sealed record ProductData(int Id, string Name, string Type, double Price);
+    public sealed record ProductData(int Id, string Name, string Type, decimal Price);
 }
