+++ Backend-Test/src/BackendTest.Domain/Entities/Product.cs
@@ -2,7 +2,7 @@ namespace BackendTest.Domain.Entities
 {
     public sealed class Product
     {
-        public Product(int id, string name, string type, double price)
+        public Product(int id, string name, string type, decimal price)
         {
             Id = id;
             Name = name;
@@ -13,6 +13,6 @@ namespace BackendTest.Domain.Entities
         public int Id { get; }
         public string Name { get; }
         public string Type { get; }
-        public double Price { get; }
+        public decimal Price { get; }
     }
 }
